+++ src/main/kotlin/step4/racingcar/domain/NumberGenerator.kt
@@ -1,4 +1,4 @@
-package step4.racingcar.racingcar.domain
+package step4.racingcar.domain
 
 interface NumberGenerator {
     fun generate(): Int
