+++ src/test/kotlin/step4/racingcar/racingcar/CarRacingTest.kt
@@ -2,8 +2,8 @@ package step4.racingcar.racingcar
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
-import step4.racingcar.racingcar.domain.Car
-import step4.racingcar.racingcar.domain.CarRacing
+import step4.racingcar.domain.Car
+import step4.racingcar.domain.CarRacing
 
 class CarRacingTest {
