+++ src/test/kotlin/step4/racingcar/racingcar/CarTest.kt
@@ -4,7 +4,7 @@ import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.assertThrows
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.ValueSource
-import step4.racingcar.racingcar.domain.Car
+import step4.racingcar.domain.Car
 
 class CarTest {
