+++ src/test/kotlin/step4/racingcar/racingcar/GameResultTest.kt
@@ -4,8 +4,8 @@ import org.assertj.core.api.Assertions
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.ValueSource
-import step4.racingcar.racingcar.domain.Car
-import step4.racingcar.racingcar.domain.GameResult
+import step4.racingcar.domain.Car
+import step4.racingcar.domain.GameResult
 
 class GameResultTest {
