+++ src/test/kotlin/step4/racingcar/racingcar/MockNumberGenerator.kt
@@ -1,6 +1,6 @@
 package step4.racingcar.racingcar
 
-import step4.racingcar.racingcar.domain.NumberGenerator
+import step4.racingcar.domain.NumberGenerator
 
 object MockNumberGenerator : NumberGenerator {
     override fun generate(): Int {
